reorganize db entities
